+++ js-plugins/vue/src/farm-vue-plugin.ts
@@ -1,27 +1,60 @@
 import fs from 'fs';
-import path from 'path';
 import { parse } from '@vue/compiler-sfc';
 import { JsPlugin } from '@farmfe/core';
 import { handleHmr } from './farm-vue-hmr.js';
-import { StylesCodeCache, CacheDescriptor, LessStatic } from './farm-vue-types.js';
+import {
+  StylesCodeCache,
+  CacheDescriptor,
+  PreProcessors,
+  PreProcessorsType,
+  PreProcessorsOptions,
+  ValueOf,
+} from './farm-vue-types.js';
 import { genMainCode } from './generatorCode.js';
-import { error } from './utils.js';
+import {
+  callWithErrorHandle,
+  error,
+  isLess,
+  isSass,
+  isStyl,
+  loadPreProcessor,
+} from './utils.js';
 
-//apply style langs
-type ApplyStyleLangs = ['less'];
+// apply style langs
+type ApplyStyleLangs = ['less', 'sass', 'scss', 'stylus'];
 
 const stylesCodeCache: StylesCodeCache = {};
-const applyStyleLangs = ['less'];
-const VueRegExp = /.vue$/;
-const JsOrTsExp = /.(ts|js)$/;
+const applyStyleLangs = ['less', 'sass', 'scss', 'stylus'];
 const cacheDescriptor: CacheDescriptor = {};
 
 export default function farmVuePlugin(options: object = {}): JsPlugin {
-  //options hooks to get farmConfig
+  // options hooks to get farmConfig
   let farmConfig = null;
   return {
     name: 'farm-vue-plugin',
     load: {
+      filters: {
+        resolvedPaths: ['.vue$'],
+      },
+      async executor(params, ctx) {
+        const { resolvedPath } = params;
+        let source = '';
+        try {
+          source = await fs.promises.readFile(resolvedPath, 'utf-8');
+        } catch (err) {
+          error({
+            id: resolvedPath,
+            message: "path is not right,can't readFile",
+          });
+        }
+        return {
+          content: source,
+          moduleType: 'ts',
+        };
+      },
+    },
+    // add hmr code In root file
+    transform: {
       filters: {
         resolvedPaths: ['.vue$'],
       },
@@ -31,38 +64,29 @@ export default function farmVuePlugin(options: object = {}): JsPlugin {
           query[key] = value;
         });
         const { vue, lang, hash } = query;
-        const { resolvedPath } = params;
-        const extname = path.extname(resolvedPath);
-        //handle .vue file
-        if (VueRegExp.test(extname)) {
-          if (vue === 'true' && hash) {
-            let styleCode = stylesCodeCache[hash];
-            //if lang is not "css",use preProcessor to handle
-            if (applyStyleLangs.includes(lang)) {
-              const { css } = await preProcession(styleCode, lang);
-              styleCode = css;
-            }
-            return {
-              content: typeof styleCode === 'string' ? styleCode : '',
-              moduleType: 'css',
-            };
+        const { resolvedPath, content: source } = params;
+        // handle .vue file
+        if (vue === 'true' && hash) {
+          let styleCode = stylesCodeCache[hash];
+          // if lang is not "css", use preProcessor to handle
+          if (applyStyleLangs.includes(lang)) {
+            const { css } = await preProcession(styleCode, lang);
+            styleCode = css;
           }
-          let source = '';
-          try {
-            source = await fs.promises.readFile(resolvedPath, 'utf-8');
-          } catch (err) {
-            error({
-              id: resolvedPath,
-              message: "path is not right,can't readFile",
-            });
-          }
-          try {
-            parse(source);
-          } catch (e) {
-            console.log(e);
-          }
-          const { descriptor } = parse(source);
+          return {
+            content: typeof styleCode === 'string' ? styleCode : '',
+            moduleType: 'css',
+          };
+        }
 
+        // transform vue
+        const result = callWithErrorHandle<null, typeof parse, [string]>(
+          this,
+          parse,
+          [source]
+        );
+        if (result) {
+          const { descriptor } = result;
           const isHmr = handleHmr(
             cacheDescriptor,
             descriptor,
@@ -71,20 +95,25 @@ export default function farmVuePlugin(options: object = {}): JsPlugin {
             resolvedPath
           );
           if (isHmr)
-            return { content: isHmr.source, moduleType: isHmr.moduleType };
+            return {
+              content: isHmr.source,
+              moduleType: isHmr.moduleType,
+              sourceMap: isHmr.map,
+            };
 
-          const { source: mainCode, moduleType } = genMainCode(
-            descriptor,
-            stylesCodeCache,
-            resolvedPath
-          );
-
+          const {
+            source: mainCode,
+            moduleType,
+            map,
+          } = genMainCode(descriptor, stylesCodeCache, resolvedPath);
           return {
             content: mainCode,
             moduleType,
+            sourceMap: map,
           };
         }
-        //default
+
+        // default
         else {
           console.error(
             `[farm-vue-plugin]:there is no path can be match,please check!`
@@ -95,14 +124,6 @@ export default function farmVuePlugin(options: object = {}): JsPlugin {
             moduleType: 'js',
           };
         }
-      },
-    },
-    // add hmr code In root file
-    transform: {
-      filters: {
-        resolvedPaths: ['.html$'],
-      },
-      executor(params, ctx) {
         return {
           content: params.content,
           moduleType: params.moduleType,
@@ -114,23 +135,80 @@ export default function farmVuePlugin(options: object = {}): JsPlugin {
 
 async function preProcession(styleCode: string, moduleType: string) {
   const __default = { css: styleCode, map: '' };
+  let processor: ValueOf<PreProcessors>;
   try {
+    // load less/sass/stylus preprocessor
+    // compile style code to css
     switch (moduleType) {
       case 'less':
-        let lessProcessor = (await import(moduleType)) || {};
-        if (lessProcessor.default) {
-          lessProcessor = lessProcessor.default;
-        }
-        return await transformLessToCss(styleCode, lessProcessor as LessStatic);
+        processor = await loadPreProcessor(PreProcessorsType.less);
+        return await compilePreProcessorCodeToCss(styleCode, processor);
+      case 'sass':
+      case 'scss':
+        processor = await loadPreProcessor(PreProcessorsType.sass);
+        return await compilePreProcessorCodeToCss(styleCode, processor, {
+          indentedSyntax: moduleType === 'sass',
+        });
+      case 'stylus':
+        processor = await loadPreProcessor(PreProcessorsType.stylus);
+        return await compilePreProcessorCodeToCss(styleCode, processor);
       default:
         return __default;
     }
   } catch (err) {
-    error({ id: 'less', message: err });
+    error({ id: moduleType, message: err });
   }
   return __default;
 }
 
-async function transformLessToCss(lessCode: string, lessProcessor: LessStatic) {
-  return await lessProcessor.render(lessCode, {});
+export async function compilePreProcessorCodeToCss<
+  T extends ValueOf<PreProcessors>
+>(
+  styleCode: string,
+  preProcessor: T,
+  options?: PreProcessorsOptions<T>
+): Promise<{ css: string }> {
+  if (isLess(preProcessor)) {
+    return await new Promise((resolve, reject) => {
+      preProcessor.render(styleCode, {}, (error, { css }) => {
+        if (error) {
+          reject(error);
+        }
+
+        resolve({ css });
+      });
+    });
+  }
+
+  if (isSass(preProcessor)) {
+    return await new Promise((resolve, reject) => {
+      preProcessor.render(
+        {
+          data: styleCode,
+          ...((options as PreProcessorsOptions<
+            PreProcessors[PreProcessorsType.sass]
+          >) ?? {}),
+        },
+        (exception, { css }) => {
+          if (exception) {
+            reject(exception);
+          }
+
+          resolve({ css: css.toString() });
+        }
+      );
+    });
+  }
+
+  if (isStyl(preProcessor)) {
+    return await new Promise((resolve, reject) => {
+      preProcessor.render(styleCode, {}, (err, css) => {
+        if (err) {
+          reject(err);
+        }
+
+        resolve({ css });
+      });
+    });
+  }
 }