+++ scripts/build.mjs
@@ -1,6 +1,7 @@
 import { execa } from 'execa';
 import { createSpinner } from 'nanospinner';
-import { resolve } from 'path';
+import { resolve, join } from 'node:path';
+import fs from 'node:fs';
 
 const DEFAULT_PACKAGE_MANAGER = 'pnpm';
 const CWD = process.cwd();
@@ -8,25 +9,28 @@ const CWD = process.cwd();
 // Build the compiler binary
 const PKG_CORE = resolve(CWD, './packages/core');
 
-// Build rust-plugin-react
-const PKG_PLUGIN_REACT = resolve(CWD, './rust-plugins/react');
+// Build rust_plugin_react
+const PKG_RUST_PLUGIN = resolve(CWD, './rust-plugins');
 
-// Build js-plugin-vue
-const PKG_PLUGIN_VUE = resolve(CWD, './js-plugins/vue');
+// Build js_plugin_path
+const PKG_JS_PLUGIN = resolve(CWD, './js-plugins');
 
+// build core command
 export const buildCore = () =>
   execa(DEFAULT_PACKAGE_MANAGER, ['build:rs'], { cwd: PKG_CORE });
 
-const plugins = [
-  execa(DEFAULT_PACKAGE_MANAGER, ['build'], { cwd: PKG_PLUGIN_REACT }),
-  execa(DEFAULT_PACKAGE_MANAGER, ['build'], { cwd: PKG_PLUGIN_VUE }),
-];
+// build rust plugins
+export const rustPlugins = () => dynamicPlugin(PKG_RUST_PLUGIN);
 
-export const buildPlugins = () => Promise.all(plugins);
+// build js plugins
+export const jsPlugins = () => dynamicPlugin(PKG_JS_PLUGIN);
 
-export const copyArtifacts = () => {
-  execa(DEFAULT_PACKAGE_MANAGER, ['copy-artifacts'], { cwd: PKG_PLUGIN_REACT });
-};
+export const buildJsPlugins = () => Promise.all(jsPlugins());
+
+export const buildRustPlugins = () => Promise.all(rustPlugins());
+
+export const copyArtifacts = () =>
+  dynamicPlugin(PKG_RUST_PLUGIN, 'copy-artifacts');
 
 export async function runTask(taskName, task) {
   const spinner = createSpinner(`Building ${taskName}`).start();
@@ -41,7 +45,8 @@ export async function runTask(taskName, task) {
 
 export async function runTaskQueue() {
   await runTask('Core', buildCore);
-  await runTask('Plugins', buildPlugins);
+  await runTask('RustPlugins', buildRustPlugins);
+  await runTask('JsPlugins', buildJsPlugins);
   await runTask('Artifacts', copyArtifacts);
 }
 
@@ -55,4 +60,14 @@ export function resolveNodeVersion() {
     console.error(`Please use Node.js v${minimumMajorVersion} or higher.`);
     process.exit(1);
   }
+}
+
+function dynamicPlugin(baseDir, command = 'build', packageManager = 'pnpm') {
+  const pluginNameMap = fs
+    .readdirSync(baseDir)
+    .filter((file) => fs.statSync(join(baseDir, file)).isDirectory());
+  const path = pluginNameMap.map((subDir) => resolve(baseDir, subDir));
+  return path.map((item) => {
+    return execa(packageManager, [command], { cwd: item });
+  });
 }