+++ js-plugins/vue/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @farmfe/js-plugin-vue
 
+## 0.2.0
+
+### Minor Changes
+
+- e09991f: Support `less/sass/scss/stylus` for style's `lang` attribute in Vue SFC
+
+## 0.1.1
+
+### Patch Changes
+
+- 3dfc64f: 1. Fix hmr does not update lazy compiled module 2. Support sourcemap for Vue SFC
+- Updated dependencies [3dfc64f]
+  - @farmfe/core@0.5.3
+
 ## 0.1.0
 
 ### Minor Changes