+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @farmfe/cli
 
+## 0.5.1
+
+### Patch Changes
+
+- 9060d9c: Add the auto installing and the check of the repeat project name
+
 ## 0.5.0
 
 ### Minor Changes